# public/openapi.v1.draft.yaml
openapi: 3.1.0
info:
  title: School Timetable & Attendance API
  version: 1.0.0-draft
  contact:
    name: Szkola API Maintainers
    url: https://szkola.tkch.eu
  description: |
    Draft of a developer-friendly API for timetable browsing, attendance tracking,
    remote approvals, and admin overrides. Uses resource-oriented endpoints,
    standard HTTP semantics, and consistent error handling.
servers:
  - url: https://api.example.com
  - url: http://localhost:8787
...
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API-Key
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      schema: { type: string }
      description: Required when using cookie authentication. Must match csrf cookie value.
...
  responses:
...
    Problem401:
      description: Unauthorized
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Problem404:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem409:
      description: Conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Problem429:
      description: Too many requests
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem500:
      description: Internal server error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807-like problem details.
      properties:
        type: { type: string, description: Unique error identifier URI or code }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Domain-specific, machine-readable code }
        fields:
          type: array
          description: Field-level validation errors
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
...
    Ref:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
      required: [id, name]
    Lesson:
      type: object
      properties:
        day: { type: string }
        lesson_num: { type: string }
        time: { type: string }
        subject: { type: string }
        teacher: { anyOf: [ { $ref: '#/components/schemas/Ref' }, { type: 'null' } ] }
        group:   { anyOf: [ { $ref: '#/components/schemas/Ref' }, { type: 'null' } ] }
        room:    { anyOf: [ { $ref: '#/components/schemas/Ref' }, { type: 'null' } ] }
      required: [day, lesson_num, time, subject]
...
    BackupEntry:
      type: object
      properties:
        filename: { type: string }
        size: { type: integer }
        mtime: { type: string, format: date-time }
      required: [filename, size, mtime]
...
paths:
...
  /v1/teachers:
    get:
      security: []
      summary: List teachers
      operationId: listTeachers
      tags: [Timetable]
      description: Returns the teacher id-to-label dictionary from timetable data.
      responses:
        '200':
          description: Map of teacher id to label
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    additionalProperties: { type: string }
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
  /v1/classes:
    get:
      security: []
      summary: List classes
      operationId: listClasses
      tags: [Timetable]
      description: Returns the class id-to-label dictionary from timetable data.
      responses:
        '200':
          description: Map of class id to label
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    additionalProperties: { type: string }
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
  /v1/rooms:
    get:
      security: []
      summary: List rooms
      operationId: listRooms
      tags: [Timetable]
      description: Returns the room id-to-label dictionary from timetable data.
      responses:
        '200':
          description: Map of room id to label
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    additionalProperties: { type: string }
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
  /v1/teachers/{id}/timetable:
    get:
      security: []
      summary: Get a teacher's timetable
      operationId: getTeacherTimetable
      tags: [Timetable]
      description: Returns lessons for a teacher by canonical id or alias.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, description: Canonical id (e.g., n12) or alias (e.g., initials like RM). Aliases may be ambiguous (409). }
      responses:
        '200':
          description: Lessons for the teacher
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, description: Canonical id resolved from alias (if alias used) }
                      lessons:
                        type: array
                        items: { $ref: '#/components/schemas/Lesson' }
                    required: [id, lessons]
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          $ref: '#/components/responses/Problem409'
  /v1/classes/{id}/timetable:
    get:
      security: []
      summary: Get a class timetable
      operationId: getClassTimetable
      tags: [Timetable]
      description: Returns class lessons by canonical id or alias, with optional group filtering.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, description: Canonical id (e.g., o5) or alias (e.g., 4TAI). Aliases may be ambiguous (409). }
        - name: group
          in: query
          required: false
          description: Group id or name (e.g., "2/2"). When provided, returns lessons for the whole class and that specific group, excluding other groups.
          schema: { type: string }
        - name: includeWhole
          in: query
          required: false
          description: Include whole-class lessons when filtering by group (default true)
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: Lessons for the class
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, description: Canonical id resolved from alias (if alias used) }
                      lessons:
                        type: array
                        items: { $ref: '#/components/schemas/Lesson' }
                    required: [id, lessons]
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          $ref: '#/components/responses/Problem409'
  /v1/rooms/{id}/timetable:
    get:
      security: []
      summary: Get a room timetable
      operationId: getRoomTimetable
      tags: [Timetable]
      description: Returns lessons scheduled in a room by canonical id or alias.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, description: Canonical id (e.g., s36) or alias (e.g., room number 407). Aliases may be ambiguous (409). }
      responses:
        '200':
          description: Lessons for the room
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string, description: Canonical id resolved from alias (if alias used) }
                      lessons:
                        type: array
                        items: { $ref: '#/components/schemas/Lesson' }
                    required: [id, lessons]
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          $ref: '#/components/responses/Problem409'
...
  /v1/refresh:
    post:
      security:
        - CookieAuth: []
      summary: Synchronously refresh timetable data using the scraper (admin-only)
      operationId: refreshTimetableSync
      tags: [Maintenance]
      description: Runs synchronous timetable refresh and returns scraper result details (admin).
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Scrape finished successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    nullable: true
                required: [ok, data]
        '409':
          description: Scraper already running
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          $ref: '#/components/responses/Problem401'
        '429':
          $ref: '#/components/responses/Problem429'
        '403':
          description: Admin required
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500':
          description: Error during scraping
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /v1/timetable/backups:
    get:
      security:
        - CookieAuth: []
      summary: List available timetable backups (admin-only)
      operationId: listTimetableBackups
      tags: [Maintenance]
      description: Lists available timetable backup files with size and timestamps (admin).
      responses:
        '200':
          description: Backups
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BackupEntry' }
                required: [ok, data]
        '403':
          description: Admin required
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          $ref: '#/components/responses/Problem401'
        '500':
          $ref: '#/components/responses/Problem500'
  /v1/timetable/restore:
    post:
      security:
        - CookieAuth: []
      summary: Restore timetable_data.json from a selected backup (admin-only)
      operationId: restoreTimetableBackup
      tags: [Maintenance]
      description: Restores timetable data from a selected backup file (admin).
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string }
              required: [filename]
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      restored: { type: boolean }
                    required: [restored]
                required: [ok, data]
        '400':
          description: Invalid filename
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Admin required
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          $ref: '#/components/responses/Problem401'
        '404':
          description: Backup not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500':
          $ref: '#/components/responses/Problem500'
...